Implement Redis Queue for comment and Request files for validating our request

+++ routes/web.php
@@ -18,31 +18,38 @@ $router->get('/', function () use ($router) {
 $router->group(['prefix' => 'v1'], function ($api) {
 
     $api->group(['middleware' => ['auth:api', 'cors']], function ($api) {
+
+        // IMPORTANT!
+        // route name (as) will be used to check user permission.
+        // permission always granted if there is no route name in the request.
+        //
+        // available permission list can be updated in config/permissions.php file.
+
         // User routes
-        $api->get('users', ['uses' => 'UserController@index']);
-        $api->get('users/{id}', ['uses' => 'UserController@show']);
-        $api->post('users', ['uses' => 'UserController@store']);
-        $api->put('users/{id}', ['uses' => 'UserController@update']);
-        $api->delete('users/{id}', ['uses' => 'UserController@destroy']);
+        $api->get('users', ['uses' => 'UserController@index', 'as' => 'users.list']);
+        $api->get('users/{id}', ['uses' => 'UserController@show', 'as' => 'users.detail']);
+        $api->post('users', ['uses' => 'UserController@store', 'as' => 'users.create']);
+        $api->put('users/{id}', ['uses' => 'UserController@update', 'as' => 'users.update']);
+        $api->delete('users/{id}', ['uses' => 'UserController@destroy', 'as' => 'users.delete']);
 
         // Role routes
-        $api->get('roles', ['uses' => 'RoleController@index']);
-        $api->get('roles/{id}', ['uses' => 'RoleController@show']);
-        $api->post('roles', ['uses' => 'RoleController@store']);
-        $api->put('roles/{id}', ['uses' => 'RoleController@update']);
-        $api->delete('roles/{id}', ['uses' => 'RoleController@destroy']);
-        $api->get('permission-list', ['uses' => 'RoleController@permissionList']);
+        $api->get('roles', ['uses' => 'RoleController@index', 'as' => 'roles.list']);
+        $api->get('roles/permission-list', ['uses' => 'RoleController@permissionList', 'as' => 'roles.permissions']);
+        $api->get('roles/{id}', ['uses' => 'RoleController@show', 'as' => 'roles.detail']);
+        $api->post('roles', ['uses' => 'RoleController@store', 'as' => 'roles.create']);
+        $api->put('roles/{id}', ['uses' => 'RoleController@update', 'as' => 'roles.update']);
+        $api->delete('roles/{id}', ['uses' => 'RoleController@destroy', 'as' => 'roles.delete']);
 
         // News routes
-        $api->get('news', ['uses' => 'NewsController@index']);
-        $api->get('news/{id}', ['uses' => 'NewsController@show']);
-        $api->post('news', ['uses' => 'NewsController@store']);
-        $api->put('news/{id}', ['uses' => 'NewsController@update']);
-        $api->delete('news/{id}', ['uses' => 'NewsController@destroy']);
+        $api->get('news', ['uses' => 'NewsController@index', 'as' => 'news.list']);
+        $api->get('news/{id}', ['uses' => 'NewsController@show', 'as' => 'news.detail']);
+        $api->post('news', ['uses' => 'NewsController@store', 'as' => 'news.create']);
+        $api->put('news/{id}', ['uses' => 'NewsController@update', 'as' => 'news.update']);
+        $api->delete('news/{id}', ['uses' => 'NewsController@destroy', 'as' => 'news.delete']);
 
-        $api->post('news/{id}/comment', ['uses' => 'NewsCommentController@store']);
-        $api->post('news/{id}/like', ['uses' => 'NewsLikeController@like']);
-        $api->delete('news/{id}/unlike', ['uses' => 'NewsLikeController@unlike']);
+        $api->post('news/{id}/comment', ['uses' => 'NewsCommentController@comment', 'as' => 'news.comment']);
+        $api->post('news/{id}/like', ['uses' => 'NewsLikeController@like', 'as' => 'news.like']);
+        $api->delete('news/{id}/unlike', ['uses' => 'NewsLikeController@unlike', 'as' => 'news.unlike']);
     });
 
 });
